core: pci: Add virtual host controller
Yes, it is that simple: For targets without physical PCI support, we can
simply expose an mmconfig region in order to provide ivshmem devices to
the guests. All we need to do is to register the configured MMIO region
but skip the mapping of the region physically. A new flag in the system
config controls this.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>